Rename shared.dart -> get_cpp_type.dart
